# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: KarisAnki API
  version: 1.0.0
  description: Shared frontend/backend contract for KarisAnki.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /decks/{deckId}/cards:
    parameters:
      - $ref: '#/components/parameters/DeckId'
    get:
      operationId: listCards
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: Card list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardList'
    post:
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Created card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/Error'
  /decks/{deckId}/cards/parse:
    parameters:
      - $ref: '#/components/parameters/DeckId'
    post:
      operationId: parseCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseCardRequest'
      responses:
        '200':
          description: Import preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '400':
          $ref: '#/components/responses/Error'
  /decks/{deckId}/cards/import:
    parameters:
      - $ref: '#/components/parameters/DeckId'
    post:
      operationId: importCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportCardsRequest'
      responses:
        '200':
          description: Import result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /decks/{deckId}/cards/pronunciation/backfill:
    parameters:
      - $ref: '#/components/parameters/DeckId'
    post:
      operationId: backfillPronunciation
      responses:
        '200':
          description: Pronunciation backfill result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PronunciationBackfillResponse'
        '404':
          $ref: '#/components/responses/Error'
  /cards/{cardId}:
    parameters:
      - $ref: '#/components/parameters/CardId'
    get:
      operationId: getCard
      responses:
        '200':
          description: Card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          $ref: '#/components/responses/Error'
    put:
      operationId: updateCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: Updated card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteCard
      responses:
        '204':
          description: Deleted
  /cards/{cardId}/reset:
    parameters:
      - $ref: '#/components/parameters/CardId'
    post:
      operationId: resetCard
      responses:
        '204':
          description: Reset
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: KARISANKI_SESSION
  parameters:
    DeckId:
      name: deckId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Query:
      name: q
      in: query
      required: false
      schema:
        type: string
    Status:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [new, review, relearn, graduated]
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          examples:
            - validation_error
            - queue_refresh
            - confirmation_required
            - rate_limited
        message:
          type: string
...
    Card:
      type: object
      required: [id, deckId, front, back, position, status, stage, relearnMode, relearnCorrectCount, stateVersion, createdAt]
      properties:
        id:
          type: integer
          format: int64
        deckId:
          type: integer
          format: int64
        front:
          type: string
        back:
          type: string
        phonetic:
          type: string
          nullable: true
        position:
          type: integer
          format: int64
        status:
          type: string
          enum: [new, review, relearn, graduated]
        stage:
          type: integer
        relearnMode:
          type: string
          enum: [NONE, FORGOT, BLURRY]
        relearnCorrectCount:
          type: integer
        dueDate:
          type: string
          format: date
          nullable: true
        stateVersion:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    CreateCardRequest:
      type: object
      required: [front]
      properties:
        front:
          type: string
          minLength: 1
        back:
          type: string
    UpdateCardRequest:
      type: object
      required: [front]
      properties:
        front:
          type: string
          minLength: 1
        back:
          type: string
    CardList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        total:
          type: integer
          format: int64
        page:
          type: integer
        pageSize:
          type: integer
...
    ParseCardRequest:
      type: object
      required: [source]
      properties:
        source:
          type: string
    ImportCardsRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          minItems: 1
          maxItems: 5000
          items:
            $ref: '#/components/schemas/ImportCardRequest'
    ImportCardRequest:
      type: object
      required: [front, back]
      properties:
        front:
          type: string
        back:
          type: string
    ImportPreview:
      type: object
      required: [items, total, validCount, duplicateCount, invalidCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ImportPreviewItem'
        total:
          type: integer
        validCount:
          type: integer
        duplicateCount:
          type: integer
        invalidCount:
          type: integer
    ImportPreviewItem:
      type: object
      required: [row, front, back, duplicate, errors]
      properties:
        row:
          type: integer
        front:
          type: string
        back:
          type: string
        duplicate:
          type: boolean
        errors:
          type: array
          items:
            type: string
    ImportResult:
      type: object
      required: [created, skippedDuplicates]
      properties:
        created:
          type: integer
        skippedDuplicates:
          type: integer
    PronunciationBackfillResponse:
      type: object
      required: [updated, unchanged, missing, notWord]
      properties:
        updated:
          type: integer
        unchanged:
          type: integer
        missing:
          type: integer
        notWord:
          type: integer
...